Better to have the bash commands in a single line
i @joao-pm-santos96 , If you have bash commands in a single line you may just click three times on top to select the complete bash command. 
If the command is mixed with teh text this is not possible, so always use a separate line for the bash commands ...

Check [rustbot](https://github.com/miguelriemoliveira/RustBot) for an exemple of a Readme.md.